t019_bufferAudio - creation of a dedicated controller for sound recording
